Pattern

+++ CalculatorCase.c
@@ -2,14 +2,14 @@
 //Using switch case making a calculator for two digits
 int main()
 {
-    int a, b, c;
+    float a, b, c;
     char ch;
 
     printf("Enter 1st number:");
-    scanf("%d", &a);
+    scanf("%f", &a);
 
     printf("Enter 2st number:");
-    scanf("%d", &b);
+    scanf("%f", &b);
 
     printf("\nA-Addition \nS-Subtraction \nM-Multiplication \nD-Division \n");
     printf("\nEnter your choice:");
@@ -19,22 +19,22 @@ int main()
     {
     case 'A': // For number: "case 1: "
         c = a + b;
-        printf("%d", c);
+        printf("%f", c);
         break;
 
     case 'S': // For number: "case 2: "
         c = a - b;
-        printf("%d", c);
+        printf("%f", c);
         break;
 
     case 'M': // For number: "case 3: "
         c = a * b;
-        printf("%d", c);
+        printf("%f", c);
         break;
 
     case 'D': // For number: "case 4: "
         c = a / b;
-        printf("%d", c);
+        printf("%f", c);
         break;
 
     default:
